+++ task1/tests/task1_tests.h
@@ -2,4 +2,4 @@
 
 #include <cstdint>
 
-void MinTest(int num_iter, int64_t max_ab, int64_t max_pq, int64_t max_lr);
+void MinTest(int num_iter, __int128_t max_ab, __int128_t max_pq, __int128_t max_lr);
